+++ Back-End/Jonas Node+MongoDB Course/Nodejs + ExpressJS/Express-API/controllers/userController.js
@@ -1,0 +1,39 @@
+//================ Get all users =========================
+exports.getAllUsers = (req, res) => {
+  res.status(503).send({
+    status: "error",
+    message: "This route is not defined yet"
+  });
+}
+
+//================ Add new users =========================
+exports.addNewUser = (req, res) => {
+  res.status(503).send({
+    status: "error",
+    message: "This route is not defined yet"
+  });
+}
+
+//================ Get single users =========================
+exports.getSingleUser = (req, res) => {
+  res.status(503).send({
+    status: "error",
+    message: "This route is not defined yet"
+  });
+}
+
+//================ Update single users =========================
+exports.updateSingleUser = (req, res) => {
+  res.status(503).send({
+    status: "error",
+    message: "This route is not defined yet"
+  });
+}
+
+//================ Delete a users =========================
+exports.deleteSingleUser = (req, res) => {
+  res.status(503).send({
+    status: "error",
+    message: "This route is not defined yet"
+  });
+}
