+++ Back-End/Jonas Node+MongoDB Course/Nodejs + ExpressJS/Express-API/index.js
@@ -20,7 +20,7 @@ app.use('/api/v1/tours', tourRoute);
 app.use('/api/v1/users', userRoute);
 
 //================= Starting the server==============
-const port = 3000;
+const port = 5000;
 
 app.listen(port, () => {
   console.log(`Server started on http://localhost:${port}`);
