+++ Back-End/Jonas Node+MongoDB Course/Nodejs + ExpressJS/Express-API/routes/userRoutes.js
@@ -1,57 +1,17 @@
 const express = require('express');
-
 const route = express.Router();
 
-//================ Get all users =========================
-const getAllUsers = (req, res) => {
-  res.status(503).send({
-    status: "error",
-    message: "This route is not defined yet"
-  });
-}
-
-//================ Add new users =========================
-const addNewUser = (req, res) => {
-  res.status(503).send({
-    status: "error",
-    message: "This route is not defined yet"
-  });
-}
-
-//================ Get single users =========================
-const getSingleUser = (req, res) => {
-  res.status(503).send({
-    status: "error",
-    message: "This route is not defined yet"
-  });
-}
-
-//================ Update single users =========================
-const updateSingleUser = (req, res) => {
-  res.status(503).send({
-    status: "error",
-    message: "This route is not defined yet"
-  });
-}
-
-//================ Delete a users =========================
-const deleteSingleUser = (req, res) => {
-  res.status(503).send({
-    status: "error",
-    message: "This route is not defined yet"
-  });
-}
-
+const userController = require('../controllers/userController');
 
 route
   .route('/')
-  .get(getAllUsers)
-  .post(addNewUser);
+  .get(userController.getAllUsers)
+  .post(userController.addNewUser);
 
 route
   .route('/:id')
-  .get(getSingleUser)
-  .patch(updateSingleUser)
-  .delete(deleteSingleUser);
+  .get(userController.getSingleUser)
+  .patch(userController.updateSingleUser)
+  .delete(userController.deleteSingleUser);
 
 module.exports = route;
